separate query and storage logic in vault
