+++ vault/src/queries.test.js
@@ -452,73 +452,6 @@ describe('src/queries.js', function () {
     })
   })
 
-  describe('getLatestEvent', function () {
-    var db
-    var getLatestEvent
-
-    beforeEach(function () {
-      db = getDatabase('test-' + uuid())
-      getLatestEvent = queries.getLatestEventWith(function () {
-        return db
-      })
-    })
-
-    afterEach(function () {
-      return db.delete()
-    })
-
-    it('returns null when no events are known', function () {
-      return getLatestEvent('account-id')
-        .then(function (result) {
-          assert.strictEqual(result, null)
-        })
-    })
-
-    it('returns the highest sorting event id for the requested account', function () {
-      return db.events.bulkAdd([
-        { accountId: 'account-a', eventId: 'a' },
-        { accountId: 'account-a', eventId: 'b' },
-        { accountId: 'account-a', eventId: 't' },
-        { accountId: 'account-a', eventId: 'c' }
-      ]).then(function () {
-        return getLatestEvent('account-a')
-          .then(function (result) {
-            assert.strictEqual(result.eventId, 't')
-          })
-      })
-    })
-  })
-
-  describe('getAllEventIds', function () {
-    var db
-    var getAllEventIds
-
-    beforeEach(function () {
-      db = getDatabase('test-' + uuid())
-      getAllEventIds = queries.getAllEventIdsWith(function () {
-        return db
-      })
-    })
-
-    afterEach(function () {
-      return db.delete()
-    })
-
-    it('returns an ordered list of all known event ids', function () {
-      return db.events.bulkAdd([
-        { accountId: 'account-a', eventId: 'a' },
-        { accountId: 'account-a', eventId: 'b' },
-        { accountId: 'account-a', eventId: 't' },
-        { accountId: 'account-a', eventId: 'c' }
-      ]).then(function () {
-        return getAllEventIds('account-a')
-          .then(function (result) {
-            assert.deepStrictEqual(result, ['a', 'b', 'c', 't'])
-          })
-      })
-    })
-  })
-
   describe('validateAndParseEvent', function () {
     it('parses referrer values into a URL', function () {
       const result = queries.validateAndParseEvent({
